# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/dashboard/permission/list": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Retrieve list of permissions",
                "operationId": "952df419b5af74fb8e7dc1f9813a1a0f",
                "parameters": [
                    {
                        "name": "key_word",
                        "in": "query",
                        "description": "Keyword to search by name or slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "slug",
                                "created_at",
                                "updated_at"
                            ],
                            "example": "id"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "View Dashboard"
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "view_dashboard"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Permission to view the dashboard"
                                            },
                                            "routes": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "dashboard"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/permission/create": {
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "Create a new permission",
                "operationId": "07d73f5ca5c9b120049f00ef6a9cc514",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "routes"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "View Dashboard"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Permission to view the dashboard"
                                    },
                                    "routes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "dashboard"
                                        }
                                    },
                                    "users": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-10-01T00:00:00Z"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "View Dashboard"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "view_dashboard"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Permission to view the dashboard"
                                        },
                                        "routes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "dashboard"
                                            }
                                        },
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "start_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-01T00:00:00Z"
                                                    },
                                                    "expires_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/permission/{id}/update": {
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "Update an existing permission",
                "operationId": "6ab72dc0361ee47e03ae86ee70358d2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "View Dashboard"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Permission to view the dashboard"
                                    },
                                    "routes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "dashboard"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "View Dashboard"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "view_dashboard"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Permission to view the dashboard"
                                        },
                                        "routes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "dashboard"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/permission/delete": {
            "delete": {
                "tags": [
                    "Permission"
                ],
                "summary": "Delete one or more permissions",
                "description": "Provide a single id or comma-separated ids to delete multiple permissions",
                "operationId": "ac3f31e671865e38ff5c6756e7b29acf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "description": "Array of permission IDs to delete",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission(s) deleted successfully"
                    }
                }
            }
        },
...
    },
...
}